requirements.txt
yfinance
pandas
win10toast; platform_system == "Windows"
plotly
alpha-vantage
google-generativeai
requests
flask==3.0.0
flask-cors==4.0.0
apscheduler==3.10.4
python-dotenv==1.0.0
